simplify the commands
