+++ src/main.ts
@@ -2,6 +2,7 @@
 import { Command } from 'commander';
 import path from 'path';
 import { checkNodeVersion } from './checkNodeVersion';
+import { authCommand } from './commands/auth';
 import { cacheCommand } from './commands/cache';
 import { configCommand } from './commands/config';
 import { deleteCommand } from './commands/delete';
@@ -68,6 +69,7 @@ async function main() {
   showCommand(program);
   versionCommand(program);
   viewCommand(program);
+  authCommand(program);
 
   const generateCommand = program.command('generate').description('Generate synthetic data');
   generateDatasetCommand(generateCommand, defaultConfig, defaultConfigPath);
